menu + contrast/brightness test
