+++ reactcode/src/config.js
@@ -1,4 +1,4 @@
-export const IPAddress = '138.231.63.90'
-export const port = '2332'
-// export const IPAddress = '0.0.0.0'
-// export const port = '8000'
+//export const IPAddress = '138.231.63.90'
+//export const port = '2332'
+export const IPAddress = 'localhost'
+export const port = '8000'
